requirements.txt
pyyaml>=5.3.1
pandas==2.2.0
beautifulsoup4==4.12.3
numpy==1.26.3
pyarrow==15.0.0
requests==2.31.0
